Transform downloaded rules into defined schema
For configured rule sets, each download script uses curl to get SARIF or
Markdown sources (in that priority) for analyzer rules and options. The
resulting files are parsed and transformed into the added analyzer-rules
schema, where any options are grouped under the one or more parent
rules.

The following rule sets are supported with this change:

* .NET Quality Rules
* .NET Style Rules
* StyleCop Rules