Generate settings data sets from downloads
Uses the downloaded rules and options to export a CSV containing details
about each entry and fields for a setting to select as well as an
optional reason.

The user is expected to populate the setting for each entry according to
their own preferences and requirements. Any new entries are
automatically added to the lists without modifying earlier selections.