Switch rule errors to warnings
Having a large number of rules prevent local debugging can be
disruptive, especially when the rule is new to the developer and they
haven't learned the muscle memory of avoiding the issue in the first
place.

Since the warnaserror compiler flag can be used to prevent warnings from
being merged into a default branch, this setting can effectively replace
the error severity.