Switch rule suggestions to warnings
The suggestion severity provides an unclear path forward for a
developer, especially when they're unfamiliar with the rule.

If a rule provides value and should be enforced, it should use the
warning severity to do so. If it does not meet this criteria, it should
not show up in a developer's IDE as something to resolve.