Add download filter by rule set
Allows a rule set name or directory to be provided to the build script.
This skips any unmatching rule set during the download process.

The settings generation steps now write a warning when no rules were
downloaded for a particular set.

The option configuration generation step writes a warning and exits when
no option settings are found. This restriction is not placed on the rule
configuration generation since rule settings must be output by at least
one set.